requirements.txt
# requirements.txt
behave
appium-python-client>=4.0.0
pytest>=7.0.0
allure-pytest>=2.12.0
pytest-bdd>=6.0.0
requests>=2.28.0
